feat: organized models, added descriptions and instructions how to launch the models server
