+++ models/relationalAgent/requirements.txt
@@ -1,0 +1,3 @@
+aiofiles
+OpenAI
+websockets
